requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
Deprecated==1.2.14
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Limiter==3.3.1
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openai==0.27.8
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
Pygments==2.15.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.4.2
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
waitress==2.1.2
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
